+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## [v0.0.17] - 2022-MM-DD
 ### API Changes
+
 ### New Features
+- `Interval Class`:
+    - `.isUnisson()`
+
 ### Improved
+- `Interval Class`:
+    - `.isAscendant()`
+    - `.isDescendant()`
+
 ### Fixed
+- `Interval Class`:
+    - `.getDiatonicInterval()`
 
 -------------------------